# internal/delivery/api/docs/openapi.yaml
openapi: 3.0.3
info:
  title: YstuPortal API
  version: 1.1.0
  description: API студенческого портала с JWT-аутентификацией через HttpOnly cookie.
servers:
  - url: /api
tags:
  - name: Auth
  - name: User
paths:
  /auth/login:
    post:
      tags: [Auth]
      summary: Создать пользовательскую сессию
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
            example:
              login: demo
              password: demo123
      responses:
        "200":
          description: JWT установлен в HttpOnly cookie
          headers:
            Set-Cookie:
              schema:
                type: string
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          description: Слишком много попыток входа
  /auth/logout:
    post:
      tags: [Auth]
      summary: Завершить сессию
      security:
        - cookieAuth: []
      responses:
        "204":
          description: Сессия завершена
        "401":
          $ref: "#/components/responses/Unauthorized"
  /user/info:
    get:
      tags: [User]
      summary: Получить безопасный профиль пользователя
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Профиль
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Пользователь не найден
  /user/grades:
    get:
      tags: [User]
      summary: Получить оценки пользователя
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Список оценок
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Grade"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          description: Upstream-сессия истекла, требуется повторный вход
  /user/admin/ping:
    get:
      tags: [User]
      summary: Проверить RBAC для роли admin
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Доступ разрешён
        "403":
          $ref: "#/components/responses/Forbidden"
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: jwt
  schemas:
    LoginRequest:
      type: object
      required: [login, password]
      additionalProperties: false
      properties:
        login:
          type: string
          minLength: 3
          maxLength: 64
        password:
          type: string
          format: password
          minLength: 6
          maxLength: 128
    User:
      type: object
      required: [first_name, last_name, username, registered, role]
      properties:
        first_name: {type: string}
        last_name: {type: string}
        patronymic: {type: string}
        username: {type: string}
        mail: {type: string, format: email}
        registered: {type: boolean}
        role: {type: string, enum: [student, teacher, admin]}
        group: {type: string}
    Grade:
      type: object
      required: [course, semester, title, type_of_control, zed, mark, evaluation, diploma]
      properties:
        course: {type: integer}
        semester: {type: integer}
        title: {type: string}
        type_of_control: {type: string}
        zed: {type: integer}
        mark: {type: string}
        evaluation: {type: string}
        diploma: {type: boolean}
        teacher: {type: string}
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: {type: string}
            message: {type: string}
        request_id: {type: string}
  responses:
    BadRequest:
      description: Некорректный запрос
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Требуется аутентификация
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Доступ запрещён
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"